chore(core): agreed pr changes
