+++ packages/nx/src/native/tui/mod.rs
@@ -4,6 +4,5 @@ pub mod components;
 pub mod config;
 pub mod lifecycle;
 pub mod pty;
-pub mod task;
 pub mod tui;
 pub mod utils;
